Add typography section
Typography is made up of three pillars:
- typographic system
- suggested sizes and styles
- proportional scaling

Explain these three pillars in the most concise way possible with examples, whilst conforming to the GDS guidelines on content style.

The suggested styles use Proxima Nova Light and Semi-Bold in conjuction with system-ui fonts to:
- Adhere to the new Barnardo's digital look and feel
- Optimise performance by minimising page weight

Apply all these to the website itself.