Add AttentionGrabbingLink component
Linking to resources is different from submitting data.

For example, you can open a link in a new window, and you can't go back once you submit data.

The implementation visually differentiates between SubmitActions and AttentionGrabbingLink, so that the user in confident in what to expect when they interact with the components.

For example, the AttentionGrabbingLink doesn't:
* include affordance
* have bold text
* use the Green colour

Rewrite SubmitAction rules to be in a consistent format with this new component.