Fix example border not showing up on iOS
Due to the proportional scaling, the border would become less than 1px on smaller viewports, and therefore not render on certain devices.

The border now uses px instead of rem to enforce the unit at all sizes.

Update the stylelint config to allow the use of px globally, while restricting the use of px on certain properties to enforce proportional scaling.