Horizontal scroll bar on Chrome OSX
As a mouse user on OSX when a mouse is plugged in a vertical scroll bar is added. Adding the vertical scroll bar makes the header overflow due to the use of `vw` in the `.Header::after` CSS:

```css
&::after {
    // ...
    margin-left: calc(50% - 50vw);
    margin-right: calc(50% - 50vw);
}
```

Proposing a very simple solution to set the `Header` `overflow-x` to hidden. This fixes the issue.